+++ lib/shared/locator.dart
@@ -13,7 +13,9 @@ import 'package:get_it/get_it.dart';
 import 'package:izmirferry/ferry/data/converters/station/station.converter.dart';
 import 'package:izmirferry/ferry/data/models/station/station.model.dart';
 import 'package:izmirferry/ferry/data/providers/schedule/schedule.provider.dart';
+import 'package:izmirferry/ferry/data/providers/station/station.provider.dart';
 import 'package:izmirferry/ferry/data/repositories/schedule/schedule.repository.dart';
+import 'package:izmirferry/ferry/data/repositories/station/station.repository.dart';
 
 void initLocator() {
   //-----//
@@ -40,6 +42,9 @@ void initLocator() {
   GetIt.I.registerLazySingleton<ScheduleProvider>(
     () => IzdenizScheduleProvider(client: GetIt.I.get()),
   );
+  GetIt.I.registerLazySingleton<StationProvider>(
+    () => IzdenizStationProvider(client: GetIt.I.get()),
+  );
 
   //------------//
   // Converters //
@@ -52,8 +57,11 @@ void initLocator() {
   // Repositories //
   //--------------//
   GetIt.I.registerLazySingleton<ScheduleRepository>(
-    () => IzdenizScheduleRepository(
-      scheduleProvider: GetIt.I.get(),
+    () => IzdenizScheduleRepository(scheduleProvider: GetIt.I.get()),
+  );
+  GetIt.I.registerLazySingleton<StationRepository>(
+    () => IzdenizStationRepository(
+      stationProvider: GetIt.I.get(),
       rawToStationConverter: GetIt.I.get(),
     ),
   );