+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### For Devs
+
+#### Changed
+
+ - Separate `getEndStations` to `StationProvider` and `StationRepository`
+ - Rename and refactor `Days` enum to `Day`
+
 ## [0.1.4] - 02.06.2023
 
 ### For Users